state: Remove old task queue impl and update all callers
