+++ state/src/applicator/task_queue.rs
@@ -47,25 +47,6 @@ fn task_not_running(task_id: TaskIdentifier) -> String {
 // | Helpers |
 // -----------
 
-/// Helper to ignore errors that come from double writes
-/// TODO(@joeykraut): Remove this once we migrate to v2
-fn double_write_helper<R>(res: std::result::Result<R, StorageError>) {
-    if let Err(e) = res {
-        warn!("error double writing: {e}");
-    }
-}
-
-/// Error message emitted when a task queue is already paused
-fn already_paused(id: TaskQueueKey) -> String {
-    format!("task queue {id} already paused")
-}
-
-/// Error message emitted when the applicator attempts to preempt a conflicting
-/// committed task
-fn already_committed(id: TaskQueueKey) -> String {
-    format!("cannot preempt committed task on queue: {id}")
-}
-
 /// Construct the running state for a newly started task
 fn new_running_state() -> QueuedTaskState {
     QueuedTaskState::Running { state: PENDING_STATE.to_string(), committed: false }
@@ -87,13 +68,11 @@ impl StateApplicator {
         let tx = self.db().new_write_tx()?;
 
         // Index the task
-        // TODO(@joeykraut): Remove v1 implementation once we migrate
-        double_write_helper(tx.add_task(&queue_key, task));
         tx.enqueue_serial_task(&queue_key, task)?;
         tx.add_assigned_task(executor, &task.id)?;
 
         // Run the task if possible
-        self.maybe_run_task(queue_key, task, &tx)?;
+        self.maybe_run_task(task, &tx)?;
         tx.commit()?;
 
         self.publish_task_updates(queue_key, task);
@@ -104,7 +83,7 @@ impl StateApplicator {
     #[instrument(skip_all, err, fields(task_id = %task_id))]
     pub fn pop_task(&self, task_id: TaskIdentifier, success: bool) -> Result<ApplicatorReturnType> {
         let tx = self.db().new_write_tx()?;
-        let keys = tx.get_queue_keys_for_task_v2(&task_id)?;
+        let keys = tx.get_queue_keys_for_task(&task_id)?;
         if keys.is_empty() {
             return Err(StateApplicatorError::reject(ERR_NO_KEY));
         }
@@ -124,7 +103,7 @@ impl StateApplicator {
 
             // If the queue is non-empty, start the next task
             if let Some(task) = tx.next_runnable_task(&key)? {
-                self.maybe_run_task(key, &task, &tx)?;
+                self.maybe_run_task(&task, &tx)?;
             }
 
             if Self::should_run_matching_engine(&executor, &key, &task, &tx)? {
@@ -147,11 +126,11 @@ impl StateApplicator {
         state: QueuedTaskState,
     ) -> Result<ApplicatorReturnType> {
         let tx = self.db().new_write_tx()?;
-        let keys = tx.get_queue_keys_for_task_v2(&task_id)?;
+        let keys = tx.get_queue_keys_for_task(&task_id)?;
 
         // Check that the task is running
         let task = tx
-            .get_task_v2(&task_id)?
+            .get_task(&task_id)?
             .ok_or_else(|| StateApplicatorError::reject(invalid_task_id(task_id)))?;
         if !task.state.is_running() {
             // This can happen if the queue was preempted by another task, and the task that
@@ -159,10 +138,8 @@ impl StateApplicator {
             return Err(StateApplicatorError::reject(task_not_running(task_id)));
         }
 
-        // TODO(@joeykraut): Remove this once we migrate to v2
-        double_write_helper(tx.transition_task(&keys[0], state.clone()));
-        tx.transition_task_v2(&task_id, state)?;
-        let updated_task = tx.get_task_v2(&task_id)?.expect("task should exist");
+        tx.transition_task(&task_id, state)?;
+        let updated_task = tx.get_task(&task_id)?.expect("task should exist");
         tx.commit()?;
 
         self.publish_task_updates_multiple(&keys, &updated_task);
@@ -176,7 +153,6 @@ impl StateApplicator {
         self.clear_task_queue(key, &tx)?;
 
         // Resume the task queue in case it was paused
-        double_write_helper(tx.resume_task_queue(&key));
         tx.commit()?;
 
         Ok(ApplicatorReturnType::None)
@@ -196,160 +172,10 @@ impl StateApplicator {
 
         // Assign the task and run it if possible
         tx.add_assigned_task(executor, &task.id)?;
-        // TODO(@joeykraut): We don't need to pass a queue key after migration
-        self.maybe_run_task(keys[0], task, &tx)?;
+        self.maybe_run_task(task, &tx)?;
 
-        // We double write to the old queue implementation to ensure consistency during
-        // the migration. However, we do not add a preemptive task to the original
-        // queue, as we don't want to "pause", we want the new locking logic to
-        // take effect
-        //
-        // TODO(@joeykraut): Remove this once the migration is complete
-        for key in keys.iter() {
-            double_write_helper(tx.add_task_front(key, task));
-        }
         tx.commit()?;
-
         self.publish_task_updates_multiple(keys, task);
-        Ok(ApplicatorReturnType::None)
-    }
-
-    /// Preempt all queues on a given task
-    ///
-    /// Multiple queues may be preempted at once to give "all or none" locking
-    /// semantics to the caller
-    ///
-    /// TODO(@joeykraut): Remove this once the migration is complete
-    #[deprecated(note = "Use `EnqueuePreemptiveTask` instead")]
-    pub fn preempt_task_queues(
-        &self,
-        keys: &[TaskQueueKey],
-        task: &QueuedTask,
-        executor: &WrappedPeerId,
-    ) -> Result<ApplicatorReturnType> {
-        let tx = self.db().new_write_tx()?;
-        for key in keys.iter() {
-            self.preempt_task_queue(*key, task, executor, &tx)?;
-        }
-
-        tx.commit()?;
-        Ok(ApplicatorReturnType::None)
-    }
-
-    /// Preempt the given task queue
-    #[instrument(skip_all, err, fields(queue_key = %key))]
-    pub fn preempt_task_queue(
-        &self,
-        key: TaskQueueKey,
-        task: &QueuedTask,
-        executor: &WrappedPeerId,
-        tx: &StateTxn<'_, RW>,
-    ) -> Result<ApplicatorReturnType> {
-        // Stop any running tasks if possible
-        let current_running_task = tx.get_current_running_task(&key)?;
-        if let Some(task) = current_running_task {
-            if task.state.is_committed() {
-                error!("cannot preempt committed task: {}", task.id);
-                let err_msg = already_committed(key);
-                return Err(StateApplicatorError::reject(err_msg));
-            }
-
-            // Otherwise transition the task to queued
-            // TODO(@joeykraut): Remove the v1 implementation once we migrate
-            let state = QueuedTaskState::Queued;
-            double_write_helper(tx.transition_task(&key, state.clone()));
-            tx.transition_task_v2(&task.id, state)?;
-        }
-
-        // If the queue is already paused, a preemptive task is already
-        // running, with which we should not conflict
-        if tx.is_queue_paused(&key)? {
-            return Err(StateApplicatorError::reject(already_paused(key)));
-        }
-
-        // Pause the queue
-        double_write_helper(tx.pause_task_queue(&key));
-        tx.add_assigned_task(executor, &task.id)?;
-        // TODO(@joeykraut): Remove the v1 implementation once we migrate
-        double_write_helper(tx.add_task_front(&key, task));
-        tx.preempt_queue_with_serial(&[key], task)?;
-        self.publish_task_updates(key, task);
-
-        Ok(ApplicatorReturnType::None)
-    }
-
-    /// Resume multiple task queues
-    pub fn resume_task_queues(
-        &self,
-        keys: &[TaskQueueKey],
-        success: bool,
-    ) -> Result<ApplicatorReturnType> {
-        if keys.is_empty() {
-            return Ok(ApplicatorReturnType::None);
-        }
-
-        let tx = self.db().new_write_tx()?;
-
-        // Resume each queue
-        for key in keys.iter() {
-            self.resume_task_queue(*key, success, &tx)?;
-        }
-        tx.commit()?;
-        Ok(ApplicatorReturnType::None)
-    }
-
-    /// Resume a task queue
-    #[instrument(skip_all, err, fields(queue_key = %key))]
-    fn resume_task_queue(
-        &self,
-        key: TaskQueueKey,
-        success: bool,
-        tx: &StateTxn<'_, RW>,
-    ) -> Result<ApplicatorReturnType> {
-        // We don't resume an unpaused queue, nor do we execute any of the dequeuing
-        // logic below, this may happen if the queue was cleared while a preemptive task
-        // was running
-        if !tx.is_queue_paused(&key)? {
-            warn!("task queue {key} is not paused, ignoring resume");
-            return Ok(ApplicatorReturnType::None);
-        }
-
-        // Resume the queue, and pop the preemptive task that was added when the queue
-        // was paused
-        tx.resume_task_queue(&key)?;
-        // TODO(@joeykraut): Remove this once we migrate to v2
-        let task_id = match tx.get_queued_tasks(&key)?.first().map(|t| t.id) {
-            Some(t) => t,
-            None => return Ok(ApplicatorReturnType::None),
-        };
-
-        // We ignore `None` here for the migration
-        // TODO(@joeykraut): Remove this once we migrate to v2
-        let (task, executor) = match self.pop_and_record_task(&key, &task_id, success, tx)? {
-            Some((t, e)) => (t, e),
-            None => return Ok(ApplicatorReturnType::None),
-        };
-
-        // If the task failed, clear the rest of the queue, as subsequent tasks will
-        // likely have invalid state
-        if !success {
-            self.clear_task_queue(key, tx)?;
-        }
-
-        // Start running the first task if it exists
-        // This will resume the task as if it is starting anew, regardless of whether
-        // the task was previously running
-        let tasks = tx.get_queued_tasks(&key)?;
-        if let Some(task) = tasks.first() {
-            self.maybe_run_task(key, task, tx)?;
-        }
-
-        // Possibly run the matching engine on the wallet
-        if Self::should_run_matching_engine(&executor, &key, &task, tx)? {
-            self.run_matching_engine_on_wallet(key, tx)?;
-        }
-
-        self.publish_task_updates(key, &task);
         Ok(ApplicatorReturnType::None)
     }
 
@@ -367,7 +193,7 @@ impl StateApplicator {
 
         // Handle in-flight tasks that were reassigned
         for task_id in reassigned_tasks.into_iter() {
-            let task = match tx.get_task_v2(&task_id)? {
+            let task = match tx.get_task(&task_id)? {
                 Some(task) => task,
                 None => continue,
             };
@@ -379,9 +205,7 @@ impl StateApplicator {
             // TODO: If the task is committed we can be smarter and check for its most
             // recent state on-chain. This is a simpler solution for the moment, but will
             // error in the case described
-            for queue in tx.get_queue_keys_for_task_v2(&task_id)? {
-                self.maybe_run_task(queue, &task, &tx)?;
-            }
+            self.maybe_run_task(&task, &tx)?;
         }
 
         tx.commit()?;
@@ -421,20 +245,13 @@ impl StateApplicator {
     }
 
     /// Transition a task into the running state
-    fn maybe_run_task(
-        &self,
-        queue_key: TaskQueueKey,
-        task: &QueuedTask,
-        tx: &StateTxn<'_, RW>,
-    ) -> Result<()> {
+    fn maybe_run_task(&self, task: &QueuedTask, tx: &StateTxn<'_, RW>) -> Result<()> {
         if !tx.can_task_run(&task.id)? {
             return Ok(());
         }
 
-        // TODO(@joeykraut): Remove v1 implementation once we migrate
         let running_state = new_running_state();
-        double_write_helper(tx.transition_task(&queue_key, running_state.clone()));
-        tx.transition_task_v2(&task.id, running_state)?;
+        tx.transition_task(&task.id, running_state)?;
         self.maybe_execute_task(task, tx)
     }
 
@@ -517,15 +334,12 @@ impl StateApplicator {
     /// Returns the task and the executor to which it was assigned
     fn pop_and_record_task(
         &self,
-        // TODO(@joeykraut): Remove this param once we migrate to v2
         key: &TaskQueueKey,
         task_id: &TaskIdentifier,
         success: bool,
         tx: &StateTxn<'_, RW>,
     ) -> Result<Option<(QueuedTask, WrappedPeerId)>> {
         // Pop the task
-        // TODO(@joeykraut): Remove this once we migrate to v2
-        double_write_helper(tx.pop_task(key));
         let mut task = match tx.pop_task_v2(task_id) {
             Ok(Some(t)) => t,
             Ok(None) => return Ok(None),
@@ -582,9 +396,7 @@ impl StateApplicator {
     /// "failed"
     fn clear_task_queue(&self, key: TaskQueueKey, tx: &StateTxn<'_, RW>) -> Result<()> {
         // Remove all tasks from queue in storage
-        // TODO(@joeykraut): Remove this once we migrate to v2
-        double_write_helper(tx.clear_task_queue(&key));
-        let cleared_tasks = tx.clear_task_queue_v2(&key)?;
+        let cleared_tasks = tx.clear_task_queue(&key)?;
 
         // Mark all tasks as failed, append to history, and publish updates
         for mut task in cleared_tasks {
@@ -630,10 +442,7 @@ impl StateApplicator {
 mod test {
     use common::types::{
         gossip::{mocks::mock_peer, WrappedPeerId},
-        tasks::{
-            mocks::{mock_preemptive_task, mock_queued_task},
-            QueuedTaskState, TaskIdentifier, TaskQueueKey,
-        },
+        tasks::{mocks::mock_queued_task, QueuedTaskState, TaskIdentifier, TaskQueueKey},
         wallet::WalletIdentifier,
         wallet_mocks::mock_empty_wallet,
     };
@@ -648,7 +457,7 @@ mod test {
         },
         storage::{
             db::DB,
-            tx::task_queuev2::{TaskQueue, TaskQueuePreemptionState},
+            tx::task_queue::{TaskQueue, TaskQueuePreemptionState},
         },
     };
 
@@ -693,7 +502,6 @@ mod test {
     fn enqueue_dummy_task(key: TaskQueueKey, db: &DB) -> TaskIdentifier {
         let task = mock_queued_task(key);
         let tx = db.new_write_tx().unwrap();
-        tx.add_task(&key, &task).unwrap();
         tx.enqueue_serial_task(&key, &task).unwrap();
 
         // Add an assignment
@@ -721,20 +529,19 @@ mod test {
 
         // Check the task was added to the queue
         let tx = applicator.db().new_read_tx()?;
-        let tasks = tx.get_queued_tasks(&task_queue_key)?;
-        tx.commit()?;
+        let queue = tx.get_task_queue(&task_queue_key)?;
+        let task_retrieved = tx.get_task(&task_id)?.unwrap();
 
-        assert_eq!(tasks.len(), 1);
-        assert_eq!(tasks[0].id, task_id);
+        let expected_queue = TaskQueue { serial_tasks: vec![task_id], ..Default::default() };
+        assert_eq!(queue, expected_queue);
         assert_eq!(
-            tasks[0].state,
+            task_retrieved.state,
             QueuedTaskState::Running { state: PENDING_STATE.to_string(), committed: false },
         ); // should be started
 
         // Check the task was started
         assert!(!task_recv.is_empty());
         let task = task_recv.recv()?;
-
         assert!(matches!(task, TaskDriverJob::Run(queued_task) if queued_task.id == task_id));
         Ok(())
     }
@@ -751,10 +558,9 @@ mod test {
 
         // Check the task was not started
         let tx = applicator.db().new_read_tx()?;
-        let tasks = tx.get_queued_tasks(&task_queue_key)?;
-        tx.commit()?;
-
-        assert_eq!(tasks.len(), 1);
+        let queue = tx.get_task_queue(&task_queue_key)?;
+        let expected_queue = TaskQueue { serial_tasks: vec![task.id], ..Default::default() };
+        assert_eq!(queue, expected_queue);
         assert!(task_recv.is_empty());
         Ok(())
     }
@@ -767,7 +573,7 @@ mod test {
         let task_queue_key = TaskQueueKey::new_v4();
 
         // Add a task directly via the db
-        enqueue_dummy_task(task_queue_key, applicator.db());
+        let task1_id = enqueue_dummy_task(task_queue_key, applicator.db());
 
         // Add another task via the applicator
         let task2 = mock_queued_task(task_queue_key);
@@ -775,12 +581,12 @@ mod test {
 
         // Ensure that the second task is in the db's queue, not marked as running
         let tx = applicator.db().new_read_tx()?;
-        let tasks = tx.get_queued_tasks(&task_queue_key)?;
-        tx.commit()?;
-
-        assert_eq!(tasks.len(), 2);
-        assert_eq!(tasks[1].id, task2.id);
-        assert_eq!(tasks[1].state, QueuedTaskState::Queued);
+        let queue = tx.get_task_queue(&task_queue_key)?;
+        let task2_retrieved = tx.get_task(&task2.id)?.unwrap();
+        let expected_queue =
+            TaskQueue { serial_tasks: vec![task1_id, task2.id], ..Default::default() };
+        assert_eq!(queue, expected_queue);
+        assert_eq!(task2_retrieved.state, QueuedTaskState::Queued);
 
         // Ensure that the task queue is empty (no task is marked as running)
         assert!(task_recv.is_empty());
@@ -804,10 +610,9 @@ mod test {
 
         // Ensure the task was removed from the queue
         let tx = applicator.db().new_read_tx()?;
-        let tasks = tx.get_queued_tasks(&wallet_id)?;
-        tx.commit()?;
-
-        assert_eq!(tasks.len(), 0);
+        let queue = tx.get_task_queue(&wallet_id)?;
+        let expected_queue = TaskQueue::default();
+        assert_eq!(queue, expected_queue);
 
         // Ensure no task was started
         assert!(task_recv.is_empty());
@@ -834,13 +639,12 @@ mod test {
 
         // Ensure the first task was removed from the queue
         let tx = applicator.db().new_read_tx()?;
-        let tasks = tx.get_queued_tasks(&task_queue_key)?;
-        tx.commit()?;
-
-        assert_eq!(tasks.len(), 1);
-        assert_eq!(tasks[0].id, task2.id);
+        let queue = tx.get_task_queue(&task_queue_key)?;
+        let task2_retrieved = tx.get_task(&task2.id)?.unwrap();
+        let expected_queue = TaskQueue { serial_tasks: vec![task2.id], ..Default::default() };
+        assert_eq!(queue, expected_queue);
         assert_eq!(
-            tasks[0].state,
+            task2_retrieved.state,
             QueuedTaskState::Running { state: PENDING_STATE.to_string(), committed: false },
         ); // should be started
 
@@ -869,20 +673,18 @@ mod test {
 
         // Ensure the first task was removed from the queue
         let tx = applicator.db().new_read_tx()?;
-        let tasks = tx.get_queued_tasks(&task_queue_key)?;
-        tx.commit()?;
-
-        assert_eq!(tasks.len(), 1);
-        assert_eq!(tasks[0].id, task2.id);
+        let queue = tx.get_task_queue(&task_queue_key)?;
+        let task2_retrieved = tx.get_task(&task2.id)?.unwrap();
+        let expected_queue = TaskQueue { serial_tasks: vec![task2.id], ..Default::default() };
+        assert_eq!(queue, expected_queue);
         assert_eq!(
-            tasks[0].state,
+            task2_retrieved.state,
             QueuedTaskState::Running { state: PENDING_STATE.to_string(), committed: false },
         ); // should be started
 
         // Ensure the second task was started
         assert!(!task_recv.is_empty());
         let task = task_recv.recv()?;
-
         assert!(matches!(task, TaskDriverJob::Run(queued_task) if queued_task.id == task2.id));
         Ok(())
     }
@@ -960,7 +762,7 @@ mod test {
 
         // Ensure the task state was updated
         let tx = applicator.db().new_read_tx()?;
-        let task_info = tx.get_task_v2(&task.id)?.unwrap();
+        let task_info = tx.get_task(&task.id)?.unwrap();
         tx.commit()?;
 
         assert_eq!(task_info.state, new_state);
@@ -1000,12 +802,7 @@ mod test {
     #[test]
     #[allow(non_snake_case)]
     fn test_clear_queue__empty() -> Result<()> {
-        let (task_queue, _task_recv) = new_task_driver_queue();
-        let applicator = mock_applicator_with_task_queue(task_queue);
-
-        // Set the local peer ID
-        let peer_id = mock_peer().peer_id;
-        set_local_peer_id(&peer_id, applicator.db());
+        let applicator = setup_mock_applicator();
 
         // Clear the queue
         let key = WalletIdentifier::new_v4();
@@ -1013,12 +810,9 @@ mod test {
 
         // Check that the queue is still empty and unpaused
         let tx = applicator.db().new_read_tx()?;
-        let is_paused = tx.is_queue_paused(&key)?;
-        let tasks = tx.get_queued_tasks(&key)?;
-        tx.commit()?;
-
-        assert!(!is_paused);
-        assert_eq!(tasks.len(), 0);
+        let queue = tx.get_task_queue(&key)?;
+        let expected_queue = TaskQueue::default();
+        assert_eq!(queue, expected_queue);
         Ok(())
     }
 
@@ -1026,14 +820,8 @@ mod test {
     #[test]
     #[allow(non_snake_case)]
     fn test_clear_queue__non_empty() -> Result<()> {
-        let (task_queue, _task_recv) = new_task_driver_queue();
-        let applicator = mock_applicator_with_task_queue(task_queue);
-
-        // Set the local peer ID
-        let peer_id = mock_peer().peer_id;
-        set_local_peer_id(&peer_id, applicator.db());
-
         // Add a task directly via the db
+        let applicator = setup_mock_applicator();
         let task_queue_key = TaskQueueKey::new_v4();
         enqueue_dummy_task(task_queue_key, applicator.db());
 
@@ -1042,12 +830,9 @@ mod test {
 
         // Check that the queue is empty and unpaused
         let tx = applicator.db().new_read_tx()?;
-        let is_paused = tx.is_queue_paused(&task_queue_key)?;
-        let tasks = tx.get_queued_tasks(&task_queue_key)?;
-        tx.commit()?;
-
-        assert!(!is_paused);
-        assert_eq!(tasks.len(), 0);
+        let queue = tx.get_task_queue(&task_queue_key)?;
+        let expected_queue = TaskQueue::default();
+        assert_eq!(queue, expected_queue);
         Ok(())
     }
 
@@ -1078,45 +863,7 @@ mod test {
         // Check the task queue
         let tx = applicator.db().new_read_tx()?;
         let task_queue = tx.get_task_queue(&task_queue_key)?;
-        tx.commit()?;
-
         assert_eq!(task_queue, TaskQueue::default());
-        Ok(())
-    }
-
-    /// Tests clearing a queue when it is paused
-    ///
-    /// TODO(@joeykraut): Remove this test once we've migrated to the new task
-    /// queue
-    #[test]
-    #[allow(non_snake_case)]
-    fn test_clear_queue__paused() -> Result<()> {
-        let (task_queue, _task_recv) = new_task_driver_queue();
-        let applicator = mock_applicator_with_task_queue(task_queue);
-
-        // Set the local peer ID
-        let peer_id = mock_peer().peer_id;
-        set_local_peer_id(&peer_id, applicator.db());
-
-        // Add a task directly via the db
-        let task_queue_key = TaskQueueKey::new_v4();
-        enqueue_dummy_task(task_queue_key, applicator.db());
-
-        // Pause the queue
-        let preemptive_task = mock_preemptive_task(task_queue_key);
-        applicator.preempt_task_queues(&[task_queue_key], &preemptive_task, &peer_id)?;
-
-        // Clear the queue
-        applicator.clear_queue(task_queue_key)?;
-
-        // Check that the queue is empty and unpaused
-        let tx = applicator.db().new_read_tx()?;
-        let is_paused = tx.is_queue_paused(&task_queue_key)?;
-        let tasks = tx.get_queued_tasks(&task_queue_key)?;
-        tx.commit()?;
-
-        assert!(!is_paused);
-        assert_eq!(tasks.len(), 0);
         Ok(())
     }
 
@@ -1156,7 +903,7 @@ mod test {
         assert_eq!(task_queue, expected_queue);
 
         // The existing task should be in the queued state
-        let existing_task = tx.get_task_v2(&serial_task.id)?.unwrap();
+        let existing_task = tx.get_task(&serial_task.id)?.unwrap();
         assert_eq!(existing_task.state, QueuedTaskState::Queued);
 
         // Lastly, the applicator should have sent the task to the driver
@@ -1201,8 +948,8 @@ mod test {
         assert_eq!(task_queue, expected_queue);
 
         // The existing tasks should be in the queued state
-        let task1_retrieved = tx.get_task_v2(&task1.id)?.unwrap();
-        let task2_retrieved = tx.get_task_v2(&task2.id)?.unwrap();
+        let task1_retrieved = tx.get_task(&task1.id)?.unwrap();
+        let task2_retrieved = tx.get_task(&task2.id)?.unwrap();
         assert!(matches!(task1_retrieved.state, QueuedTaskState::Running { .. }));
         assert!(matches!(task2_retrieved.state, QueuedTaskState::Running { .. }));
 
@@ -1343,7 +1090,7 @@ mod test {
 
         // Check the task status of the original task
         let tx = applicator.db().new_read_tx()?;
-        let task1_retrieved = tx.get_task_v2(&task1.id)?.unwrap();
+        let task1_retrieved = tx.get_task(&task1.id)?.unwrap();
         assert!(matches!(task1_retrieved.state, QueuedTaskState::Running { .. }));
 
         let task_queue1 = tx.get_task_queue(&queue_key1)?;
@@ -1393,7 +1140,7 @@ mod test {
 
         // Check the task status of the original task
         let tx = applicator.db().new_read_tx()?;
-        let task2_retrieved = tx.get_task_v2(&task2.id)?.unwrap();
+        let task2_retrieved = tx.get_task(&task2.id)?.unwrap();
         assert!(matches!(task2_retrieved.state, QueuedTaskState::Running { .. }));
 
         let task_queue1 = tx.get_task_queue(&queue_key1)?;
@@ -1405,152 +1152,6 @@ mod test {
         };
         assert_eq!(task_queue1, TaskQueue::default());
         assert_eq!(task_queue2, expected_queue2);
-        Ok(())
-    }
-
-    // TODO(@joeykraut): Remove the tests below once we've migrated to the new task
-    // queue implementation
-
-    /// Tests cases in pausing an empty queue
-    #[test]
-    fn test_pause_empty() -> Result<()> {
-        let (task_queue, task_recv) = new_task_driver_queue();
-        let applicator = mock_applicator_with_task_queue(task_queue);
-        let peer_id = mock_peer().peer_id;
-        set_local_peer_id(&peer_id, applicator.db());
-
-        let task_queue_key = TaskQueueKey::new_v4();
-
-        // Pause the queue
-        let preemptive_task = mock_preemptive_task(task_queue_key);
-        applicator.preempt_task_queues(&[task_queue_key], &preemptive_task, &peer_id)?;
-
-        // Ensure the queue was paused
-        let tx = applicator.db().new_read_tx()?;
-        let is_paused = tx.is_queue_paused(&task_queue_key)?;
-        tx.commit()?;
-
-        assert!(is_paused);
-
-        // Add a task and ensure it is not started
-        let task = mock_queued_task(task_queue_key);
-        let task_id = task.id;
-        applicator.append_task(&task, &peer_id)?;
-
-        let tx = applicator.db().new_read_tx()?;
-        let tasks = tx.get_queued_tasks(&task_queue_key)?;
-        tx.commit()?;
-
-        assert!(task_recv.is_empty());
-        assert_eq!(tasks.len(), 2); // Includes the preemptive task
-        assert_eq!(tasks[0].state, QueuedTaskState::Preemptive);
-        assert_eq!(tasks[1].state, QueuedTaskState::Queued);
-
-        // Resume the queue and ensure the task is started
-        applicator.resume_task_queues(&[task_queue_key], true /* success */)?;
-
-        let tx = applicator.db().new_read_tx()?;
-        let task = tx.get_current_running_task(&task_queue_key)?;
-        tx.commit()?;
-        assert_eq!(task.unwrap().id, task_id);
-
-        assert!(
-            matches!(task_recv.recv()?, TaskDriverJob::Run(queued_task) if queued_task.id == task_id)
-        );
-        Ok(())
-    }
-
-    /// Tests pausing a queue with a running task
-    #[test]
-    fn test_pause_with_running() -> Result<()> {
-        let (task_queue, task_recv) = new_task_driver_queue();
-        let applicator = mock_applicator_with_task_queue(task_queue);
-        let peer_id = mock_peer().peer_id;
-        set_local_peer_id(&peer_id, applicator.db());
-
-        let task_queue_key = TaskQueueKey::new_v4();
-
-        // Add a task
-        let task = mock_preemptive_task(task_queue_key);
-        let task_id = task.id;
-        applicator.append_task(&task, &peer_id)?;
-
-        // Start the task
-        let tx = applicator.db().new_write_tx()?;
-        let state = QueuedTaskState::Running { state: PENDING_STATE.to_string(), committed: false };
-        tx.transition_task(&task_queue_key, state)?;
-        tx.commit()?;
-
-        // Pause the queue
-        let preemptive_task = mock_preemptive_task(task_queue_key);
-        applicator.preempt_task_queues(&[task_queue_key], &preemptive_task, &peer_id)?;
-
-        // Ensure the queue was paused
-        let tx = applicator.db().new_read_tx()?;
-        let is_paused = tx.is_queue_paused(&task_queue_key)?;
-        tx.commit()?;
-
-        assert!(is_paused);
-
-        // Ensure the task was transitioned to queued
-        let tx = applicator.db().new_read_tx()?;
-        let tasks = tx.get_queued_tasks(&task_queue_key)?;
-        tx.commit()?;
-
-        assert_eq!(tasks.len(), 2); // Includes the preemptive task
-        assert_eq!(tasks[0].state, QueuedTaskState::Preemptive);
-        assert_eq!(tasks[1].state, QueuedTaskState::Queued);
-
-        // Resume the queue and ensure the task is started
-        applicator.resume_task_queues(&[task_queue_key], true /* success */)?;
-
-        let tx = applicator.db().new_read_tx()?;
-        let task = tx.get_current_running_task(&task_queue_key)?;
-        tx.commit()?;
-
-        assert!(task.is_some());
-        assert_eq!(task.unwrap().id, task_id);
-
-        assert!(!task_recv.is_empty());
-        assert!(
-            matches!(task_recv.recv()?, TaskDriverJob::Run(queued_task) if queued_task.id == task_id)
-        );
-        Ok(())
-    }
-
-    /// Test the case in which preempting one queue of multiple fails
-    #[test]
-    fn test_preempt_multiple_fail() -> Result<()> {
-        let (task_queue, _task_recv) = new_task_driver_queue();
-        let applicator = mock_applicator_with_task_queue(task_queue);
-        let peer_id = mock_peer().peer_id;
-        set_local_peer_id(&peer_id, applicator.db());
-
-        let queue_key1 = TaskQueueKey::new_v4();
-        let queue_key2 = TaskQueueKey::new_v4();
-
-        // Add a preemptive task to the first queue
-        let task = mock_preemptive_task(queue_key1);
-        applicator.preempt_task_queues(&[queue_key1], &task, &peer_id)?;
-
-        // Try to preempt both task queues
-        let new_task = mock_preemptive_task(queue_key2);
-        let result = applicator.preempt_task_queues(&[queue_key1, queue_key2], &new_task, &peer_id);
-        assert!(matches!(result, Err(StateApplicatorError::Rejected(..))));
-
-        // Verify that the task that _did_ have an existing preemptive task is still
-        // paused
-        let tx = applicator.db().new_read_tx()?;
-        let is_paused1 = tx.is_queue_paused(&queue_key1)?;
-        tx.commit()?;
-        assert!(is_paused1);
-
-        // Verify that the task queue that did not originally have a preemptive task in
-        // it remains unpaused
-        let tx = applicator.db().new_read_tx()?;
-        let is_paused2 = tx.is_queue_paused(&queue_key2)?;
-        tx.commit()?;
-        assert!(!is_paused2);
         Ok(())
     }
 
@@ -1660,144 +1261,6 @@ mod test {
         assert!(
             matches!(task_recv.recv()?, TaskDriverJob::Run(queued_task) if queued_task.id == task2.id)
         );
-        Ok(())
-    }
-
-    // TODO(@joeykraut): Remove double write tests once migration is complete
-
-    // ---------------------------------
-    // | Task Queue Double Write Tests |
-    // ---------------------------------
-
-    /// Tests that an added task is double-written correctly
-    #[test]
-    #[allow(non_snake_case)]
-    fn test_add_and_pop_task__double_write() -> Result<()> {
-        let applicator = setup_mock_applicator();
-
-        // Add a task
-        let executor = mock_peer().peer_id;
-        let task = mock_queued_task(WalletIdentifier::new_v4());
-        let queue_key = task.descriptor.queue_key();
-        applicator.append_task(&task, &executor)?;
-
-        // Check the task queue v2 storage for the task
-        let tx = applicator.db().new_read_tx()?;
-        let task_info_some = tx.get_task_v2(&task.id)?.is_some();
-        assert!(task_info_some);
-
-        let is_empty = tx.is_queue_empty_v2(&queue_key)?;
-        let task_queue = tx.get_task_queue(&queue_key)?;
-        tx.commit()?;
-        assert!(!is_empty);
-
-        // Check the task queue for the wallet
-        let expected_queue = TaskQueue {
-            serial_tasks: vec![task.id],
-            concurrent_tasks: vec![],
-            preemption_state: TaskQueuePreemptionState::NotPreempted,
-        };
-        assert_eq!(task_queue, expected_queue);
-
-        // Pop the task
-        applicator.pop_task(task.id, true /* success */)?;
-
-        // Check that the task queue v2 storage is updated
-        let tx = applicator.db().new_read_tx()?;
-        let task_info_none = tx.get_task_v2(&task.id)?.is_none();
-        assert!(task_info_none);
-
-        let is_empty = tx.is_queue_empty_v2(&queue_key)?;
-        let task_queue = tx.get_task_queue(&queue_key)?;
-        assert!(is_empty);
-        assert_eq!(task_queue, TaskQueue::default());
-        Ok(())
-    }
-
-    /// Checks double write behavior when multiple tasks are added to the queue
-    #[test]
-    #[allow(non_snake_case)]
-    fn test_add_multiple_tasks__double_write() -> Result<()> {
-        let applicator = setup_mock_applicator();
-
-        // Add two tasks
-        let executor = mock_peer().peer_id;
-        let queue_key = WalletIdentifier::new_v4();
-        let task1 = mock_queued_task(queue_key);
-        let task2 = mock_queued_task(queue_key);
-        applicator.append_task(&task1, &executor)?;
-        applicator.append_task(&task2, &executor)?;
-
-        // Check the task queue v2 storage for the tasks
-        let tx = applicator.db().new_read_tx()?;
-        let task1_info_some = tx.get_task_v2(&task1.id)?.is_some();
-        let task2_info_some = tx.get_task_v2(&task2.id)?.is_some();
-        assert!(task1_info_some);
-        assert!(task2_info_some);
-
-        let task_queue = tx.get_task_queue(&queue_key)?;
-        let expected_queue = TaskQueue {
-            serial_tasks: vec![task1.id, task2.id],
-            concurrent_tasks: vec![],
-            preemption_state: TaskQueuePreemptionState::NotPreempted,
-        };
-
-        assert_eq!(task_queue, expected_queue);
-        tx.commit()?;
-        Ok(())
-    }
-
-    /// Tests the double write behavior when a queue is paused and subsequently
-    /// resumed
-    #[test]
-    #[allow(non_snake_case)]
-    fn test_pause_and_resume_queue__double_write() -> Result<()> {
-        let applicator = setup_mock_applicator();
-
-        // Add a task
-        let executor = mock_peer().peer_id;
-        let queue_key = WalletIdentifier::new_v4();
-        let task = mock_queued_task(queue_key);
-        applicator.append_task(&task, &executor)?;
-
-        // Pause the queue
-        let preemptive_task = mock_preemptive_task(queue_key);
-        applicator.preempt_task_queues(&[queue_key], &preemptive_task, &executor)?;
-
-        // Check the task queue v2 storage for the task
-        let tx = applicator.db().new_read_tx().unwrap();
-        let task_info_some = tx.get_task_v2(&task.id)?.is_some();
-        let preemptive_task_info_some = tx.get_task_v2(&preemptive_task.id)?.is_some();
-        assert!(task_info_some);
-        assert!(preemptive_task_info_some);
-        let task_queue = tx.get_task_queue(&queue_key)?;
-        tx.commit()?;
-
-        // Check the task queue for the wallet
-        let expected_queue = TaskQueue {
-            serial_tasks: vec![preemptive_task.id, task.id],
-            concurrent_tasks: vec![],
-            preemption_state: TaskQueuePreemptionState::SerialPreemptionQueued,
-        };
-        assert_eq!(task_queue, expected_queue);
-
-        // Resume the queue, then check the task queue v2 storage for the task
-        applicator.resume_task_queues(&[queue_key], true /* success */)?;
-        let tx = applicator.db().new_read_tx()?;
-        let task_info_some = tx.get_task_v2(&task.id)?.is_some();
-        let preemptive_task_info_none = tx.get_task_v2(&preemptive_task.id)?.is_none();
-        let task_queue = tx.get_task_queue(&queue_key)?;
-        assert!(task_info_some);
-        assert!(preemptive_task_info_none);
-        tx.commit()?;
-
-        // Check the task queue for the wallet
-        let expected_queue = TaskQueue {
-            serial_tasks: vec![task.id],
-            concurrent_tasks: vec![],
-            preemption_state: TaskQueuePreemptionState::NotPreempted,
-        };
-        assert_eq!(task_queue, expected_queue);
         Ok(())
     }
 }
