+++ state/src/lib.rs
@@ -164,13 +164,6 @@ pub enum StateTransition {
     TransitionTask { task_id: TaskIdentifier, state: QueuedTaskState },
     /// Clear all tasks in the queue, marking them as failed
     ClearTaskQueue { queue: TaskQueueKey },
-    /// Preempt the given task queue
-    ///
-    /// Returns any running tasks to `Queued` state and pauses the queue
-    /// 
-    /// TODO(@joeykraut): Remove this once the migration is complete
-    #[deprecated(note="Use `EnqueuePreemptiveTask` instead")]
-    PreemptTaskQueues { keys: Vec<TaskQueueKey>, task: QueuedTask, executor: WrappedPeerId },
     /// Enqueue a preemptive task for the given task queues
     /// 
     /// Transitions any running tasks to `Queued` state and enqueues the given task
@@ -181,11 +174,6 @@ pub enum StateTransition {
     /// concurrent task and may share the queue with other concurrent, preemptive
     /// tasks.
     EnqueuePreemptiveTask { keys: Vec<TaskQueueKey>, task: QueuedTask, executor: WrappedPeerId, serial: bool },
-    /// Resume the given task queue
-    /// 
-    /// TODO(@joeykraut): Remove this once the migration is complete
-    #[deprecated(note="Use `PopTask` instead, as in the standard task flow")]
-    ResumeTaskQueues { keys: Vec<TaskQueueKey>, success: bool},
     /// Reassign all tasks from one peer to another peer
     ReassignTasks { from: WrappedPeerId, to: WrappedPeerId },
