+++ workers/task-driver/src/driver.rs
@@ -8,10 +8,7 @@ use job_types::task_driver::{TaskDriverJob, TaskDriverReceiver, TaskNotification
 use state::State;
 use tokio::runtime::Builder as TokioRuntimeBuilder;
 use tracing::{error, info, instrument, warn};
-use util::{
-    concurrency::{new_shared, Shared},
-    telemetry::helpers::backfill_trace_field,
-};
+use util::concurrency::{new_shared, Shared};
 
 use crate::{
     error::TaskDriverError,
@@ -173,9 +170,6 @@ impl TaskExecutor {
                 )
                 .await
             },
-            TaskDriverJob::RunImmediate { task_id, task, resp } => {
-                self.handle_run_immediate(task_id, task, resp).await
-            },
             TaskDriverJob::Notify { task_id, channel } => {
                 self.handle_notification_request(task_id, channel).await
             },
@@ -203,92 +197,13 @@ impl TaskExecutor {
         Ok(())
     }
 
-    /// Handle a request to run a task immediately
-    #[instrument(skip_all, err, fields(task_id = %task_id, affected_wallets))]
-    #[deprecated(note = "send preemptive tasks through the raft")]
-    async fn handle_run_immediate(
-        &self,
-        task_id: TaskIdentifier,
-        task: TaskDescriptor,
-        resp: Option<TaskNotificationSender>,
-    ) -> Result<(), TaskDriverError> {
-        let affected_wallets = task.affected_wallets();
-        backfill_trace_field("affected_wallets", format!("{affected_wallets:?}"));
-
-        // Check if any non-preemptable tasks conflict with this task before pausing
-        for wallet_id in affected_wallets.iter() {
-            if let Some(s) = self.preemption_conflict(wallet_id).await? {
-                warn!("task preemption failed: {s}");
-                if let Some(chan) = resp {
-                    let _ = chan.send(Err(s));
-                }
-
-                return Ok(());
-            }
-        }
-
-        self.start_preemptive_task(affected_wallets, task_id, task, resp).await
-    }
-
-    /// Check for any conflicting conditions that cannot be preempted
-    ///
-    /// Returns an error message describing the conflict, if one exists
-    async fn preemption_conflict(
-        &self,
-        wallet_id: &WalletIdentifier,
-    ) -> Result<Option<String>, TaskDriverError> {
-        if let Some(conflicting_task) = self.state().current_committed_task(wallet_id).await? {
-            return Ok(Some(format!(
-                "task preemption conflicts with committed task {conflicting_task:?}, aborting..."
-            )));
-        }
-
-        if self.state().is_queue_paused(wallet_id).await? {
-            return Ok(Some("queue already paused, aborting...".to_string()));
-        }
-
-        Ok(None)
-    }
-
     // ------------------
     // | Task Execution |
     // ------------------
 
-    /// Start the given task, preempting the queues of the given wallets
-    #[deprecated(note = "send preemptive tasks through the raft")]
-    async fn start_preemptive_task(
-        &self,
-        affected_wallets: Vec<WalletIdentifier>,
-        task_id: TaskIdentifier,
-        task: TaskDescriptor,
-        resp: Option<TaskNotificationSender>,
-    ) -> Result<(), TaskDriverError> {
-        // Pause the queues for the affected local wallets
-        if !affected_wallets.is_empty() {
-            let waiter = self
-                .state()
-                .pause_multiple_task_queues(affected_wallets.clone(), task_id, task.clone())
-                .await?;
-            waiter.await?;
-        }
-
-        let res = self.start_task(true /* immediate */, task_id, task, affected_wallets).await;
-        if let Err(e) = &res {
-            error!("error running immediate task: {e:?}");
-        }
-
-        // Notify the task creator that the task has completed
-        if let Some(sender) = resp {
-            let _ = sender.send(res.map_err(|e| e.to_string()));
-        }
-
-        Ok(())
-    }
-
     /// Spawn a new task in the driver
     ///
     /// Returns the success of the task
-
     #[instrument(name = "task", skip_all, err, fields(
         task_id = %id,
         task = %task.display_description(),
